# backend/openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: Jahan Academy API
  version: 1.0.0-draft
  description: Contract-first API. x-phase identifies MVP and later complete-product operations.
servers:
  - url: /api/v1
...
paths:
  /reference-data/{kind}:
    get:
      tags: [Reference Data]
      operationId: listReferenceData
      x-phase: MVP
      parameters:
        - $ref: '#/components/parameters/ReferenceKind'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - { name: countryId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: parentId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '200': { $ref: '#/components/responses/ReferenceDataPage' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/reference-data/{kind}:
    get:
      tags: [Reference Data]
      operationId: listAdminReferenceData
      x-phase: MVP
      x-permissions: [reference_data.read]
      security: [{ UserAccessToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ReferenceKind'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/ReferenceDataPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [Reference Data]
      operationId: createReferenceData
      x-phase: MVP
      x-permissions: [reference_data.write]
      security: [{ UserAccessToken: [] }]
      parameters: [{ $ref: '#/components/parameters/ReferenceKind' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReferenceDataWrite' }
      responses:
        '201': { $ref: '#/components/responses/ReferenceDataItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/reference-data/{kind}/{id}:
    parameters:
      - $ref: '#/components/parameters/ReferenceKind'
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Reference Data]
      operationId: getAdminReferenceData
      x-phase: MVP
      x-permissions: [reference_data.read]
      security: [{ UserAccessToken: [] }]
      responses:
        '200': { $ref: '#/components/responses/ReferenceDataItem' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Reference Data]
      operationId: replaceReferenceData
      x-phase: MVP
      x-permissions: [reference_data.write]
      security: [{ UserAccessToken: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReferenceDataWrite' }
      responses:
        '200': { $ref: '#/components/responses/ReferenceDataItem' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Reference Data]
      operationId: archiveReferenceData
      x-phase: MVP
      x-permissions: [reference_data.write]
      security: [{ UserAccessToken: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '200': { $ref: '#/components/responses/ReferenceDataItem' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
  /countries:
    get:
      tags: [Public Catalog]
      operationId: listCountries
      x-phase: MVP
      parameters:
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: featured
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Published countries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CountrySummary' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /countries/{slug}:
    get:
      tags: [Public Catalog]
      operationId: getCountry
      x-phase: MVP
      parameters:
        - $ref: '#/components/parameters/Slug'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200':
          description: Published country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryEnvelope'
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    StaffSession:
      type: apiKey
      in: cookie
      name: jahan_staff_session
    UserAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshToken:
      type: apiKey
      in: cookie
      name: jahan_refresh
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string, pattern: '^(W/)?"[1-9][0-9]*"$' }
    ReferenceKind:
      name: kind
      in: path
      required: true
      schema:
        type: string
        enum: [countries, cities, academic-levels, fields-of-study, intakes, currencies]
    Id:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 160, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
    Locale:
      name: locale
      in: query
      schema: { type: string, enum: [fa, en], default: fa }
    Search:
      name: q
      in: query
      schema: { type: string, minLength: 2, maxLength: 100 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    PreconditionFailed:
      description: The supplied version is stale; reload the item before retrying
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ReferenceDataPage:
      description: Localized reference-data page
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ReferenceDataPage' }
    ReferenceDataItem:
      description: Reference-data item including both translations for staff callers
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { $ref: '#/components/schemas/ReferenceDataItem' }
...
    Unauthorized:
      description: Authentication required or invalid
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Forbidden:
      description: Authenticated but not permitted
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Resource does not exist or is not visible to this caller
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Conflict:
      description: Resource or idempotency conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    ValidationError:
      description: Request failed field or domain validation
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    ReferenceTranslation:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 180 }
        description: { type: [string, 'null'], maxLength: 5000 }
    ReferenceTranslations:
      type: object
      additionalProperties: false
      required: [fa, en]
      properties:
        fa: { $ref: '#/components/schemas/ReferenceTranslation' }
        en: { $ref: '#/components/schemas/ReferenceTranslation' }
    ReferenceDataWrite:
      type: object
      description: Kind-specific write payload; unknown fields are rejected and fa/en are mandatory.
      required: [translations]
      properties:
        translations: { $ref: '#/components/schemas/ReferenceTranslations' }
        code: { type: string }
        iso2: { type: string, pattern: '^[A-Za-z]{2}$' }
        iso3: { type: [string, 'null'], pattern: '^[A-Za-z]{3}$' }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        countryId: { $ref: '#/components/schemas/Uuid' }
        parentId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        status: { type: string, enum: [draft, published, archived] }
        active: { type: boolean }
        featured: { type: boolean }
        displayOrder: { type: integer, minimum: 0, maximum: 1000000 }
        numericCode: { type: [string, 'null'], pattern: '^\d{3}$' }
        symbol: { type: string, minLength: 1, maxLength: 12 }
        decimalPlaces: { type: integer, minimum: 0, maximum: 4 }
    ReferenceDataItem:
      type: object
      required: [id, type, code, name, active, displayOrder, version]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        type: { type: string }
        code: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        active: { type: boolean }
        displayOrder: { type: integer }
        version: { type: integer, minimum: 1 }
        slug: { type: [string, 'null'] }
        countryId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        parentId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        status: { type: [string, 'null'] }
        translations: { $ref: '#/components/schemas/ReferenceTranslations' }
    ReferenceDataPage:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ReferenceDataItem' }
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
    Uuid:
      type: string
      format: uuid
    Meta:
      type: object
      additionalProperties: false
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    PageEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: {} }
        meta: { $ref: '#/components/schemas/Meta' }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, fieldErrors, requestId]
          properties:
            code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
            message: { type: string }
            fieldErrors:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
            requestId: { type: string }
    CountrySummary:
      type: object
      additionalProperties: false
      required: [id, slug, code, name, featured]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        slug: { type: string }
        code: { type: string, pattern: '^[A-Z]{2}$' }
        name: { type: string }
        featured: { type: boolean }
        imageUrl: { type: [string, 'null'], format: uri }
    CountryEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/CountrySummary' }
...